third_party/blink/web_tests/external/wpt/webaudio/the-audio-api/the-audiocontext-interface/audiocontext-sinkid-setsinkid.https-expected.txt

This is a testharness.js-based test.
PASS setSinkId() with a valid device identifier should succeeded.
PASS setSinkId() with the same sink ID should resolve immediately.
FAIL setSinkId() with the same AudioSinkOptions.type value should resolve immediately. assert_equals: expected "object" but got "string"
FAIL setSinkId() should fail with TypeError on an invalid AudioSinkOptions.type value. assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL setSinkId() should fail with NotFoundError on an invalid device identifier. assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
